by_month outputs
The output for by_month was customers - seems a bit off since we were putting in merchants.

Not sure if this is what you were thinking - also the spec is still a bit unclear on when you say "most_popular_merchants" I couldn't find this method else where in the project. (Most popular by revenue in said month? Most popular by invoice in said month?)

Thanks again for working on this - Allan